- refactor : board 파일 위치 pages 폴더로 이동
