+++ StudentSoup/src/main/frontend/src/pages/board/board.scss
