+++ StudentSoup/src/main/frontend/src/pages/board/boardDetail.scss
