+++ StudentSoup/src/main/frontend/src/pages/board/boardcomponents/MobileBoard.tsx
