+++ StudentSoup/src/main/frontend/src/pages/board/boardcomponents/PcBoard.tsx
